refactor(scripts): SMI-4783 extract release helpers under 500-line budget
prepare-release.ts: 794 → 403 lines via extraction to scripts/lib/:
  - release-collision.ts (304 lines): npm-view, collision rules, RESERVED_RANGES
  - release-changelog.ts (54 lines): findLastVersionBumpCommit + prependToChangelog
  - release-git.ts (89 lines): validatePostWrite + getCurrentBranch + createCommit

Public surface preserved verbatim — prepare-release.ts re-exports the helper
symbols so scripts/tests/prepare-release.test.ts and any other consumers
continue to import from '../prepare-release'.

Unblocks SMI-4775 (lockfile regen step), which previously could not be added
without bumping the file past the pre-commit check-file-length.mjs hard limit.
